backend/requirements.txt
flask
authlib
python-dotenv
flask-sqlalchemy
pymysql
cryptography
requests
flask-cors
flasgger
python-jose
